requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.6.1
email-validator==2.1.0
motor==3.6.0
beanie==1.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
authlib==1.3.2
itsdangerous==2.2.0
cloudinary==1.41.0
